refactor: rename component, correct trad and add export
