+++ packages/apps/helpdesk/src/components/templates/index.js
@@ -19,12 +19,12 @@
 export {default as ContactPartnerSearchBar} from './ContactPartnerSearchBar/ContactPartnerSearchBar';
 export {default as CustomerSearchBar} from './CustomerSearchBar/CustomerSearchBar';
 export {default as MyTeamTicketSearchBar} from './MyTeamTicketSearchBar/MyTeamTicketSearchBar';
+export {default as MyTicketSearchBar} from './MyTicketSearchBar/MyTicketSearchBar';
 export {default as ProjectSearchBar} from './ProjectSearchBar/ProjectSearchBar';
 export {default as TicketDropdownCards} from './TicketDropdownCards/TicketDropdownCards';
 export {default as TicketEditButton} from './TicketEditButton/TicketEditButton';
 export {default as TicketHeader} from './TicketHeader/TicketHeader';
 export {default as TicketPriorityPicker} from './TicketPriorityPicker/TicketPriorityPicker';
-export {default as TicketSearchBar} from './TicketSearchBar/TicketSearchBar';
 export {default as TicketsStatusButton} from './TicketsStatusButton/TicketsStatusButton';
 export {default as TicketStopwatch} from './TicketStopwatch/TicketStopwatch';
 export {default as TicketTypeSearchBar} from './TicketTypeSearchBar/TicketTypeSearchBar';
